refactor: extract VIEW tests into ShareLinkBucketViewTest with tearDown cleanup
Move 10 VIEW tests from ShareLinkBucketTest into a dedicated
ShareLinkBucketViewTest class that uses setUp/tearDown for automatic
cleanup instead of verbose try/finally blocks in each test method.
